% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{HyperParameters}
\alias{HyperParameters}
\title{Set hyperparamters used in the model}
\usage{
HyperParameters(
  object.list,
  d = 35,
  eta = NULL,
  mu_mu = NULL,
  Sigma_mu = NULL,
  nu_omega = NULL,
  n_Lambda = NULL,
  Sigma_Lambda = NULL,
  nu_tau = NULL,
  lambda0 = NULL,
  lambda1 = NULL,
  alpha_p = NULL,
  beta_p = NULL
)
}
\arguments{
\item{object.list}{list of seurat object}

\item{d}{dimension of latent factor.}

\item{eta}{smoothing parameters in Potts model.}

\item{mu_mu}{mean vector of gaussian distribution (mu).}

\item{Sigma_mu}{shared covariance matrix of gaussian distribution (mu).}

\item{nu_omega}{shape and rate parameters of gamma distribution (omega).}

\item{n_Lambda}{degree of freedom of wishart distribution (Lambda).}

\item{Sigma_Lambda}{pds location matrix (Lambda).}

\item{nu_tau}{shape and rate parameters of gamma distribution (tau).}

\item{lambda0}{variance of gaussian distribution (L).}

\item{lambda1}{variance of gaussian distribution (L).}

\item{alpha_p}{shape1 parameter of beta distribution (p).}

\item{beta_p}{shape2 parameter of beta distribution (p).}
}
\value{
result: a list of hyperparameters
}
\description{
Set hyperparamters used in the model
}
