% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{InitiParameters}
\alias{InitiParameters}
\title{Initializing parameters used in EM algorithm later}
\usage{
InitiParameters(object.list, k = 7, d = 35)
}
\arguments{
\item{object.list}{a list of interested seurat object (preprocessed).}

\item{k}{number of subtypes.}

\item{d}{dimension of latent space.}
}
\value{
a list of initialized parameters.
}
\description{
Initializing parameters used in EM algorithm later
}
