% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_findMNN.R
\name{mnn_adjacent}
\alias{mnn_adjacent}
\title{Find mutual nearest neighbors pairs}
\usage{
mnn_adjacent(
  batch.list,
  d = 50,
  BSPARAM = ExactParam(),
  BPPARAM = SerialParam()
)
}
\arguments{
\item{batch.list}{list of gene expression.}

\item{d}{dimension of PCA.}

\item{BSPARAM}{parallel.}

\item{BPPARAM}{parallel.}
}
\value{
a sparse matrix - adjacent matrix
}
\description{
Find mutual nearest neighbors pairs
}
