% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{PreprocessData}
\alias{PreprocessData}
\title{Functions for preprocessing data}
\usage{
PreprocessData(
  object.list,
  min.features = 200,
  min.spots = 20,
  nfeatures = 2000
)
}
\arguments{
\item{object.list}{A list of Seurat objects.}

\item{min.features}{Filtering spots no more than min.features expression gene.}

\item{min.spots}{Filter genes no more than min.spots expression spots.}

\item{nfeatures}{Number of high variable features.}
}
\value{
a list of Seurat objects
}
\description{
Functions for preprocessing data
}
