% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadia-package.R
\docType{package}
\name{stadia-package}
\alias{stadia-package}
\alias{_PACKAGE}
\title{stadia: Statistical Batch-Aware Embedded Integration, Dimension Reduction and Alignment for Spatial Transcriptomics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Integrate multi-source spatial transcriptomics data sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yanfang-li/stadia}
  \item \url{https://yanfang-li.github.io/stadia/}
  \item Report bugs at \url{https://github.com/yanfang-li/stadia/issues}
}

}
\author{
\strong{Maintainer}: Yanfang Li \email{liyanfang1110@gmail.com}

}
\keyword{internal}
