% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stadia}
\alias{stadia}
\title{Core function to conduct STADIA algorithm}
\usage{
stadia(
  object.list,
  hyper,
  dim = 35,
  n_cluster = 7,
  platform = c("visium", "st", "others"),
  adj.cutoff = 50,
  icm.maxiter = 10,
  em.maxiter = 30,
  min.features = 200,
  min.spots = 20,
  nfeatures = 2000,
  verbose = TRUE,
  verbose.in = TRUE,
  ncores = NULL
)
}
\arguments{
\item{object.list}{a list of seurat object.}

\item{hyper}{a list of all hyperParameters.}

\item{dim}{dimension of hidden space.}

\item{n_cluster}{number of spatial domains.}

\item{platform}{string vector to idicate platform generating data}

\item{adj.cutoff}{cutoff when calculating adjacent matrix.}

\item{icm.maxiter}{number of iteration of icm.}

\item{em.maxiter}{number of iteration of em.}

\item{min.features}{filtering low-quality data.}

\item{min.spots}{filtering low-quality data.}

\item{nfeatures}{number of HVGs.}

\item{verbose}{print information or not.}

\item{verbose.in}{print information or not.}

\item{ncores}{number of cores to use.}
}
\value{
a list of estimated parameters.
}
\description{
core function to conduct STADIA algorithm
}
